hello

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -1,13 +1,14 @@
 #include <stdio.h>
+#include <math.h>
 /**
  *main - main point
  *Return: always 0
  */
 int main(void)
 {
-	long long int n = 612852475143;
-	long long int largest_factor = 1;
-	long long int i;
+	long int n = 612852475143;
+	long int largest_factor = 1;
+	long int i;
 
 	for (i = 2; i <= n; i++)
 	{
@@ -17,6 +18,6 @@ int main(void)
 			n = n / i;
 		}
 	}
-	printf("%lld\n", largest_factor);
+	printf("%ld\n", largest_factor);
 	return (0);
 }
